+++ Dyalect/Runtime/Types/DyCollection.cs
@@ -119,24 +119,6 @@ namespace Dyalect.Runtime.Types
 
         internal override int GetCount() => Count;
 
-        public override object ChangeType(Type type)
-        {
-            if (type == Dyalect.Types.ArrayObject)
-                return ConvertToArray();
-            else if (type.IsArray)
-            {
-                var et = type.GetElementType();
-                var arr = Array.CreateInstance(et, Count);
-
-                for (var i = 0; i < arr.Length; i++)
-                    arr.SetValue(GetValue(i).ChangeType(et), i);
-
-                return arr;
-            }
-            else
-                return base.ChangeType(type);
-        }
-
         internal DyObject[] Concat(ExecutionContext ctx, DyObject right)
         {
             var newArr = new List<DyObject>(GetValues());