+++ Dyalect/Runtime/Types/DyObject.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 
 namespace Dyalect.Runtime.Types
 {
@@ -94,31 +95,6 @@ namespace Dyalect.Runtime.Types
         public override sealed bool Equals(object obj) => obj is DyObject dyo && Equals(dyo);
 
         public override int GetHashCode() => HashCode.Combine(TypeId);
-
-        public virtual object ChangeType(Type type)
-        {
-            if (type == Dyalect.Types.Boolean)
-                return GetBool();
-            else if (type == Dyalect.Types.Char)
-                return GetChar();
-            else if (type == Dyalect.Types.String)
-                return GetString();
-            else if (type == Dyalect.Types.Int32)
-                return (int)GetInteger();
-            else if (type == Dyalect.Types.Int64)
-                return GetInteger();
-            else if (type == Dyalect.Types.Double)
-                return GetFloat();
-            else if (type == Dyalect.Types.Float)
-                return (float)GetFloat();
-            else if (type == Dyalect.Types.Object)
-                return ToObject();
-            //else if (type == Dyalect.Types.ArrayObject 
-            //    && this is IEnumerable<DyObject> seq)
-
-
-            throw new InvalidCastException();
-        }
     }
 
     internal static class DyObjectInternalExtensions